+++ src/main/java/lotto/view/input/PurchaseAmountInputView.java
@@ -1,9 +1,10 @@
-package lotto.view;
+package lotto.view.input;
 
 import camp.nextstep.edu.missionutils.Console;
 import lotto.util.NumberValidator;
 
 public class PurchaseAmountInputView {
+    private static final NumberValidator numberValidator = new NumberValidator();
     private static final String PURCHASE_AMOUNT_MESSAGE = "구입금액을 입력해 주세요.";
 
     public int getInput() {
@@ -12,7 +13,7 @@ public class PurchaseAmountInputView {
             String amount = Console.readLine();
 
             try {
-                NumberValidator.validatePurchaseAmount(amount);
+                numberValidator.validatePurchaseAmount(amount);
                 return Integer.parseInt(amount);
             } catch (IllegalArgumentException e) {
                 System.out.println(e.getMessage());
